Creates panimg module
