+++ app/grandchallenge/cases/image_builders/dicom.py
@@ -7,8 +7,8 @@ import SimpleITK
 import numpy as np
 import pydicom
 
-from grandchallenge.cases.image_builders.types import ImageBuilderResult
 from grandchallenge.cases.image_builders.utils import convert_itk_to_internal
+from panimg.types import ImageBuilderResult
 
 NUMPY_IMAGE_TYPES = {
     "character": SimpleITK.sitkUInt8,
