+++ app/grandchallenge/cases/image_builders/nifti.py
@@ -3,8 +3,8 @@ from typing import Set
 
 import SimpleITK
 
-from grandchallenge.cases.image_builders.types import ImageBuilderResult
 from grandchallenge.cases.image_builders.utils import convert_itk_to_internal
+from panimg.types import ImageBuilderResult
 
 
 def format_error(message):
