+++ setup.cfg
@@ -4,6 +4,7 @@ application-import-names =
     config
     grandchallenge
     tests
+    panimg
 import-order-style = pycharm
 docstring-convention = numpy
 max-complexity = 10
